removed some more warnings and did some refactoring
