+++ src/main/java/com/mg105/Main.java
@@ -9,8 +9,8 @@ import com.mg105.interface_adapters.MapGeneratorInterpreter;
 import com.mg105.presenters.InventoryPresenter;
 import com.mg105.use_cases.Inventory.InventoryInteractor;
 import com.mg105.use_cases.MapGenerator;
+import com.mg105.user_interface.InventoryDisplay;
 import com.mg105.user_interface.MapGeneratorButton;
-import com.mg105.user_interface.inventory.InventoryDisplay;
 import com.mg105.utils.PartyConstants;
 import javafx.application.Application;
 import javafx.scene.Scene;
